requirements.txt
opencv-python==4.12.0
tensorflow==2.17.0
keras==2.17.0
numpy==1.25.2
streamlit==1.27.0
Pillow==10.1.0
h5py==3.9.0
deepface==0.0.93
